Add framework for multiple representations to be displayed in same render view.
New class to take care of the images in the stack. Reworked the framework for pqwindowimage class to support adding images and pointing to functions in the vlvastackimage class. Single image works, multiple images seem to not appear on the view. Log scaling doesn't work -- need to track down why not. Need to add subset support to additional images -- figure out how to return a value from the subsetselector UI widget.